Clean header file include
