+++ src/ShareLogParser.h
@@ -25,23 +25,15 @@
 #define SHARELOGPARSER_H_
 
 
-#include "Kafka.h"
 #include "MySQLConnection.h"
 #include "Statistics.h"
 #include "zlibstream/zstr.hpp"
 
-#include "bitcoin/CommonBitcoin.h"
-#include "bitcoin/StratumBitcoin.h"
 
 #include <event2/event.h>
 #include <event2/http.h>
 #include <event2/buffer.h>
-#include <event2/util.h>
 #include <event2/keyvalq_struct.h>
-
-#include <string.h>
-#include <pthread.h>
-#include <memory>
 
 
 ///////////////////////////////  ShareLogDumper  ///////////////////////////////
