+++ src/Statistics.h
@@ -25,19 +25,8 @@
 #define STATISTICS_H_
 
 #include "Common.h"
-#include "Kafka.h"
-#include "Stratum.h"
 
-#include <event2/event.h>
-#include <event2/http.h>
-#include <event2/buffer.h>
-#include <event2/util.h>
-#include <event2/keyvalq_struct.h>
-
-#include <string.h>
-#include <pthread.h>
-#include <memory>
-
+#include "glog/logging.h"
 
 ////////////////////////////////// StatsWindow /////////////////////////////////
 // none thread safe
