+++ src/StatsHttpd.inl
@@ -22,8 +22,14 @@
  THE SOFTWARE.
  */
 
+#include "utilities_js.hpp"
+
 #include <boost/algorithm/string.hpp>
 #include <boost/thread.hpp>
+
+#include <event2/http.h>
+#include <event2/buffer.h>
+#include <event2/keyvalq_struct.h>
 
 ////////////////////////////////  WorkerShares  ////////////////////////////////
 template <class SHARE>
