+++ src/StratumServer.h
@@ -26,26 +26,17 @@
 
 #include "Common.h"
 
-#include <map>
-#include <unordered_map>
-#include <vector>
-#include <memory>
-#include <bitset>
-#include <atomic>
-#include <fstream>
-
-#include <arpa/inet.h>
-#include <sys/socket.h>
-
-#include <event2/bufferevent.h>
-#include <event2/buffer.h>
-#include <event2/listener.h>
-#include <event2/util.h>
-#include <event2/event.h>
-#include <glog/logging.h>
-
 #include "Kafka.h"
 #include "Stratum.h"
+
+#include <bitset>
+
+#include <event2/bufferevent.h>
+#include <event2/listener.h>
+#include <event2/event.h>
+
+#include <glog/logging.h>
+
 
 class Server;
 class StratumJobEx;
