+++ src/StratumSession.h
@@ -26,18 +26,15 @@
 
 #include "Common.h"
 
-#include <netinet/in.h>
-#include <deque>
-#include <unordered_map>
-#include <event2/event.h>
-#include <event2/buffer.h>
-#include <event2/bufferevent.h>
-
-#include <glog/logging.h>
-
 #include "utilities_js.hpp"
 #include "Stratum.h"
 #include "Statistics.h"
+
+#include <deque>
+
+#include <event2/bufferevent.h>
+
+#include <glog/logging.h>
 
 
 #define CMD_MAGIC_NUMBER      0x7Fu
