+++ src/eth/BlockMakerEth.cc
@@ -24,6 +24,8 @@
 #include "BlockMakerEth.h"
 #include "EthConsensus.h"
 
+#include "utilities_js.hpp"
+
 #include <boost/thread.hpp>
 
 ////////////////////////////////////////////////BlockMakerEth////////////////////////////////////////////////////////////////
